+++ process/tests/pipeline.rs
@@ -1,18 +1,21 @@
-use pimalaya_process::Cmd;
+use pimalaya_process::{Cmd, Error};
 
 #[tokio::test]
 async fn pipeline() {
     env_logger::builder().is_test(true).init();
 
     let cmd = Cmd::from(vec!["echo hello", "cat"]);
-    let res = cmd.run().await.unwrap();
+    let out = cmd.run().await.unwrap().to_string_lossy();
 
-    assert_eq!(res.code, 0);
-    assert_eq!(res.read_out_lossy(), "hello\n");
+    assert_eq!(out, "hello\n");
 
     let cmd = Cmd::from(vec!["echo hello", "bad", "cat"]);
-    let res = cmd.run().await.unwrap();
-
-    assert_eq!(res.code, 127);
-    assert_eq!(res.read_out_lossy(), "sh: line 1: bad: command not found\n");
+    match cmd.run().await.unwrap_err() {
+        Error::InvalidExitStatusCodeNonZeroError(cmd, status, err) => {
+            assert_eq!(cmd, "bad");
+            assert_eq!(status, 127);
+            assert_eq!(err, "sh: line 1: bad: command not found\n");
+        }
+        err => panic!("unexpected error: {err:?}"),
+    }
 }