+++ process/tests/single_cmd.rs
@@ -1,21 +1,19 @@
-use pimalaya_process::Cmd;
+use pimalaya_process::{Cmd, Error};
 
 #[tokio::test]
 async fn single_cmd() {
     env_logger::builder().is_test(true).init();
 
-    let cmd = Cmd::from("echo <msg> <msg>").replace("<msg>", "hello");
-    let res = cmd.run().await.unwrap();
+    let cmd = Cmd::from("echo <msg>1 <msg>2").replace("<msg>", "hello");
+    let out = cmd.run().await.unwrap().to_string_lossy();
+    assert_eq!(out, "hello1 hello2\n");
 
-    assert_eq!(res.code, 0);
-    assert_eq!(res.read_out().unwrap(), "hello hello\n");
-
-    let cmd = Cmd::from("bad");
-    let res = cmd.run().await.unwrap();
-
-    assert_eq!(res.code, 127);
-    assert_eq!(
-        res.read_out().unwrap(),
-        "sh: line 1: bad: command not found\n"
-    );
+    match Cmd::from("bad").run().await.unwrap_err() {
+        Error::InvalidExitStatusCodeNonZeroError(cmd, status, err) => {
+            assert_eq!(cmd, "bad");
+            assert_eq!(status, 127);
+            assert_eq!(err, "sh: line 1: bad: command not found\n");
+        }
+        err => panic!("unexpected error: {err:?}"),
+    }
 }